Move graph utils methods to graph_utils.py
